Addition of a new test
This is the test from matlab.
By Looking carefully, the result is slightly different because the sign of ynew[1] is opposite
to the sign computed in matlab. This, after investigation, comes from a difference in the computation of pvalue & mvalue on approximately the
fiftheen significant digit, which leads to mvalue < pvalue in python instead of mvalue > pvalue in matlab. It is interesting to keep this in
memory, but this is not really a bug.