Modification on compare_matlabarray
-One minor modification converts the matlabarray to python arrays such that we do not have to
increment the indices anymore after flatening

-One other modification is to set the relative indices to zero where absolutes values at thoses
indices are inferior to the absolute tolerance. This makes sense to me since when numbers are inferior
to a certain number (1e-15 for instance) it does not make sense to compare the relative values since
there may be numerical errors due to zero approximation at this point, errors due to number in the machines.
This indeed made one test in replace_in_Y fail.