Addition of a test for bcdfo_poisdness which revealed a few errors
The precedent test written made not really sense I think, it had hardcons=0 which is the variable
to say that there are no constraints and then created some constraints.

This made me also notice the fact that the file bcdfo_find_new_yj_bc is not translated yet. I also only noticed then that I am missing this matlab
file, which exists though on Dave's data. However having this file or not does not seem to affect the results on the tests problems, though they have bounded constraints.

I am therefore notifying it as a bug/fix but I will not translate it immediately without further information on it.