Fixed bugs on ecdfo_check_convex
The previous tests did not reveal the bugs. Bugs were revealed when checking the finals
eigenvalues which turned out to be non real. Then I compared to the matlab code and fixed
the error, which came from 2 differents things : eig_ returns first the eigenvalues and then
the eigenvectors, which is the reverse order in matlab. Then, the eigenvalues are  returned
as a vector and not as a diagonal matrix like in matlab.