Lots of modification in comments on translation
There are a lot of problems which we do not have to care about anymore normally after modification
of the runtime.py function.
I put these problems at the end in this because there are no guarantees that this is now fixed so they should not be forgotten.
But they are no more main priorities:

Otherwise, a few others issues were added, but there are also minors since thoses are issues
told during the executing by the console, so that we can not miss them.