IMPORTANT modification on the runtime.
Following modification of the __getitem__ function, a bug was created since
__setitem__ uses __getitem__. I therefore had to modify the code in __setitem__.
Moreover, I added a case that I forgot. Until now, I only transposed matrices such as
V[m:n,i] where i is an integer. But what I also do now, which had to be done, is doing the same thing
for V[m:n,a:a], where a:a is a slice  containing a unique integer. The corresponding comment
has been modified in Comments_on_translation.

An other minor modification is the addition of qr_ function.