Important modification in the runtime
I already dealt with the special case M[:,i] where 'i'  is an integer or an interval containing a unique number but
there was another case which needed to be dealt with : V[m:n] where V is a vertical vector. This also
gave back a line vector when it should give a column vector. This is normally now dealt with
by modifying again __getitem__ and __setitem__.