Modification in the runtime
I now realized on one test what the case I had commented was for. So I uncommented it and fixed the issue I had on another case with an try/except. This may not be the best correction, but it is working and if there are no further issues coming from thoses lines it will be kept that way.