Correction of ecdfo_solve_TR_bc and his tests
There were errors in the tests, the tests with almostEqual told true for any value, and correctlm was false.
I also found a different xnew but not that much different.
However, in both cases, xnew has only a relative precision (between matlab and python) of 1e-3. Not sure if it is a bug or normal.

Otherwise, almost only modifications that I wrote in Comments_on_translation, so this is perfect.