Modification on ecdfo_augmX_evalf
The was an error in the fprintf of the code, which said row vector instead of column vector.

Also, there is there no need to transpose. In matlab it works with or without transpose,
but in python we have to be careful : we have to give it a  column vector and not a row vector.