Modifications in ecdfo_main_test
Mainly 2 indices are permuted during argsort computation jsorted=matlabarray(argsort_(np.asarray(d).reshape(-1))[::-1])+1 in bcdfo_repair_Y

The important outputs should not be modified, but columns have to be permuted at some places.

Also some times precision are not very high, but this is what could be expected when computing a lot and accumulating small differences with the matlab execution, but nothing really worrying.