-Suppression of 'char' matrices everywhere and a few modifications on almost every disp_ and fprintf_ -Also modification of the fprintf_ function in the runtime -And addition on Comments_on_translation on the modifications that have to be done on disp_ and fprintf_
This took quite a lot of time, and there are still a few errors but I think these modifications are the way to go.
Thoses 'char' matrices are not really adapted here, python has a very nice way of dealing with strings and therefore I think that
having strings instead will be easier.
Those transformations are not ended yet, more modifications about this are to come.