Modification of runtime.py and addition of tests in runtime_test.py
This inversion is important, otherwise the following test does not work :
        V=matlabarray([[1]])
        V[:,2]=2
        self.assertTrue(isequal_(V, [[1,2]]))

Other than that, I verified some other cases with other tests. What is also important to notice, is that adding rows due to the F-Contiguous property of matlabarrays do not work. Fortunately, our code does not seem to use that. Otherwise, we would have to figure out a method to deal with it in runtime or convert every case where it happens with some concatenate