Moved All files and tests in Almost_operational_but_slow_python_code
The name of the directory is quite long but I tried to make it explicit, and I quote
what I have written in the README file to explain this move :

This is a python implementation of the SQPDFO Optimizer.
This is working quite fine (there are at some points differences with the matlab implentation which are normals, since built-in functions do not always are the same results,
and may even change with the version of python/matlab used), though there no doubts remains a few bugs (since we can not simply make exhaustive tests) but thoses remaining bugs should be easy to fix with the file
Comments_on_translation or a little bit of modifications in the runtime.py (hopefully no difficult bugs remain)
Unfortunately, implementing it so with matlabarrrays make it so that some big problems, when functions are called thousands of times, are running very slow.


Conclusion : I will try to implement it using direct python arrays, and I will do modifications on copys of those already almost_operational files.