Correction of a bug with SVD decomposition
According to numpy documentation (for version 1.9.2, which I am currently using), it gives
U,S,V such that u * np.diag(s) * v. Hence the transpose of V to be coherent with the matlab code,
which returns U,S,V such that u * s * v.H. But after tests, V returned in matlab and in python
are the same...so I deleted the transpose.

Also, there was still a basic scalar multiplication in bcdfo_build_QR_of_Y though we want a
matrix multiplication, which was another error.