Correction of dimensions in bcdfo_find_new_yj, bcdfo_include_in_Y, bcdfo_repair_Y and ecdfo_find_smallf
This concerns the difference in dimensions returned by Y[:,[0]] and Y[:,0].
One should be very careful with it, because python replicates arrays to match dimensions, and then we obtain something we do not want at all,
and it takes time debugging. A comment has been add on Comments_on_translation, which will be commit later with other comments.